# Conflicts:
#	电信运营商云管平台工单处理子模块1.0/mainshow/src/components/leader/pendTickets.vue

+++ 电信运营商云管平台工单处理子模块1.0/back2.0/src/main/java/com/example/back2/entity/table/Department.java
@@ -17,7 +17,7 @@ public class Department implements Serializable {
     
     private Integer depLevel;
     
-    private Integer depBudget;
+    private Double depBudget;
 
 
     public Integer getDepNum() {
@@ -44,11 +44,11 @@ public class Department implements Serializable {
         this.depLevel = depLevel;
     }
 
-    public Integer getDepBudget() {
+    public Double getDepBudget() {
         return depBudget;
     }
 
-    public void setDepBudget(Integer depBudget) {
+    public void setDepBudget(Double depBudget) {
         this.depBudget = depBudget;
     }
